Extract Parsely class to its own file
It’s good practice to keep classes defined in their own file, with no other globally-scoped code present. This allows the class to be included, without side-effects.
